Split action_codec in two files: actions_codec and codec
